Fix infinite blocking issue
For some requests, in particular very large ones, do_cgi() blocked
indefinitely.
The problem was:

stderr.read_to_end(err_output).await.unwrap_or(0);
stdout.read_to_end(&mut data).await.unwrap_or(0);

The CGI script is not guaranteed to close it's handles in any
particular order but this code requires it to first close stderr,
before it will go and read from stdout.
In the case of a very large response the internal buffer for stdout
would fill up causing the CGI script to blocked and never reaching
the point where it would close stderr (because it will only close
stderr once it has completed writing all output to stdout, which it
can't because nobody is reading and the buffer is full).

The solution is to read stderr and stdout concurrently and waiting
for both to complete with try_join!()

This also adds an optionally compiled binary for testing and adjusts
the version number.